Refactor out fields from BurningManCandidate which are not affecting data for delayed payout tx receivers. This allows to split off the burn target related code to separate classes.
