+++ desktop/src/main/java/bisq/desktop/main/dao/burnbsq/burningmen/BurnOutputListItem.java
@@ -37,7 +37,8 @@ class BurnOutputListItem {
     BurnOutputListItem(BurnOutputModel model, BsqFormatter bsqFormatter) {
 
         height = model.getHeight();
-        cycleIndex = model.getCycleIndex();
+        // We show first cycle with index 0 as 1
+        cycleIndex = model.getCycleIndex() + 1;
         date = model.getDate();
         dateAsString = DisplayUtils.formatDateTime(new Date(date));
         amount = model.getAmount();
